+++ test/pipeline_test.dart
@@ -133,7 +133,7 @@ class PipelineInitWithGet extends Module {
 
 class RVPipelineModule extends Module {
   RVPipelineModule(Logic a, Logic reset, Logic validIn, Logic readyForOut,
-      {bool testingAsyncReset = false})
+      {bool testingAsyncReset = false, dynamic aResetVal})
       : super(name: 'rv_pipeline_module') {
     final clk = testingAsyncReset ? Const(0) : SimpleClockGenerator(10).clk;
     a = addInput('a', a, width: a.width);
@@ -149,7 +149,8 @@ class RVPipelineModule extends Module {
           (p) => [p.get(a) < p.get(a) + 1],
           (p) => [p.get(a) < p.get(a) + 1],
         ],
-        asyncReset: testingAsyncReset);
+        asyncReset: testingAsyncReset,
+        resetValues: aResetVal != null ? {a: aResetVal} : null);
     b <= pipeline.get(a);
 
     addOutput('validOut') <= pipeline.validPipeOut;
@@ -205,7 +206,6 @@ void main() {
       final pipem = SimplePipelineModule(Logic(width: 8));
       await pipem.build();
 
-      //TODO: test that reset value works!
       final vectors = [
         Vector({'a': 1}, {}),
         Vector({'a': 2}, {}),
@@ -390,6 +390,38 @@ void main() {
       final vectors = [
         Vector({'reset': 0, 'a': 1, 'validIn': 0, 'readyForOut': 1}, {}),
         Vector({'reset': 1}, {'validOut': 0}),
+      ];
+      await SimCompare.checkFunctionalVector(pipem, vectors);
+      SimCompare.checkIverilogVector(pipem, vectors);
+    });
+
+    test('rv pipeline simple reset vals', () async {
+      final pipem = RVPipelineModule(Logic(width: 8), Logic(), Logic(), Logic(),
+          aResetVal: 5);
+      await pipem.build();
+
+      final vectors = [
+        Vector({'reset': 1, 'a': 1, 'validIn': 0, 'readyForOut': 1}, {}),
+        Vector({'reset': 1, 'a': 1, 'validIn': 0, 'readyForOut': 1}, {}),
+        Vector({'reset': 1, 'a': 1, 'validIn': 0, 'readyForOut': 1}, {}),
+        Vector({'reset': 0, 'a': 1, 'validIn': 1, 'readyForOut': 1},
+            {'validOut': 0, 'b': 5}),
+        Vector({'reset': 0, 'a': 2, 'validIn': 1, 'readyForOut': 1},
+            {'validOut': 0, 'b': 6}),
+        Vector({'reset': 0, 'a': 3, 'validIn': 1, 'readyForOut': 1},
+            {'validOut': 0, 'b': 7}),
+        Vector({'reset': 0, 'a': 4, 'validIn': 1, 'readyForOut': 1},
+            {'validOut': 1, 'b': 4}),
+        Vector({'reset': 0, 'a': 0, 'validIn': 0, 'readyForOut': 1},
+            {'validOut': 1, 'b': 5}),
+        Vector({'reset': 0, 'a': 0, 'validIn': 0, 'readyForOut': 1},
+            {'validOut': 1, 'b': 6}),
+        Vector({'reset': 0, 'a': 0, 'validIn': 0, 'readyForOut': 1},
+            {'validOut': 1, 'b': 7}),
+        Vector({'reset': 0, 'a': 0, 'validIn': 0, 'readyForOut': 1},
+            {'validOut': 0}),
+        Vector({'reset': 0, 'a': 0, 'validIn': 0, 'readyForOut': 1},
+            {'validOut': 0}),
       ];
       await SimCompare.checkFunctionalVector(pipem, vectors);
       SimCompare.checkIverilogVector(pipem, vectors);